code/demo_application/requirements.txt
pandas
together
python-dotenv
jinja2
flask